refactor: 모듈 이름 변경
