+++ frontend/src/components/@layout/PageLoadingWithLogo/PageLoadingWithLogo.stories.tsx
